Fix raycast lens distortion offset
If we are raycasting a 3D transform onto a surface, our first step is to
get the screen-space position of the input transform.

The input transform is the marker node, which (if there is a lens model
node attached to the camera) will be automatically undistorted.

Therefore, it seems that we don't need to perform any lens
re/un-distortion at all.