Core Solver - Relationships from DAG path names.
We add (yet another) relationship analyis method to mmSolver.

This one can produce very fast results compared to he dependency-graph
analyis, but makes some assumptions that are fast, but may lead to
inefficent solving.

Overall, I think this is a good trade-off and this new method is a good
contender for becoming the new default algorithm.