Move build dependencies to mmsolverlibs
This moves the cminpack, Eigen, glog, ceres and OpenMVG to build with
mmsolverlibs - not mmsolver.

This allows us to keep Maya depending only on the things that it really
uses and tries to put as much non-Maya specific logic into mmsolverlibs.

This has involved adding wrappers around cminpack and OpenMVG.
Ultimately, this should (in theory) allow us to reimplement these
functions as we wish without requiring to actually depend on cminpack or
OpenMVG.