Build - Merge mmSolverLibs with mmSolver
Because the build times are much faster (12 mins down to 7 mins), and
the the overall setup is easier.

Also, there was no real use for mmSolverLibs anyway. We will continue to
seperate the dependencies and code so that mmSolverLibs is Maya-free,
but trying to seperate these into different projects was causing more
problems that it fixed.

This also fixes the incorrect installation of header files and static
libraries.