mmscenegraph - Refine and update curve fit tests and charts.
We add a lot of new tests, showing the different functionallity of the
curve_fit.rs file.

The curve_fit.rs implementation has also been simplified a little,
inlining a "multi-stage" function, and simplifying logic.

For the generated test charts, there is now a legend in the top-right
corner, with labels.

The bounds for many of the charts were also wrong and didn't take into
account all data points, which has now been fixed.